refactor(fe): move utils file to _libs folder
